[rrc] split toolchain resolution and the link steps into driver::link
The rustc/lld/libdir discovery, the polyffi paths, the scratch archive
members, the bundled launcher source, and link_product — the driver's
own link invocation — move together.
